Renamed okta login screen to okta redirect login screen
