+++ login-workflow/example/src/screens/OktaRedirectLogin.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { OktaLoginScreen } from '@brightlayer-ui/react-auth-workflow';
+import { OktaRedirectLoginScreen } from '@brightlayer-ui/react-auth-workflow';
 import EatonLogo from '../assets/images/eaton_stacked_logo.png';
 import { DebugComponent } from '../components/DebugComponent';
 
 export const OktaLogin = (): JSX.Element => (
-    <OktaLoginScreen
+    <OktaRedirectLoginScreen
         projectImage={<img src={EatonLogo} alt="logo" style={{ maxHeight: 80 }} />}
         header={<DebugComponent />}
         errorDisplayConfig={{
