--- login-workflow/src/screens/OktaLoginScreen/index.ts
@@ -1,2 +1,0 @@
-export * from './OktaLoginScreenBase';
-export * from './OktaLoginScreen';
