+++ login-workflow/src/screens/OktaRedirectLoginScreen/OktaRedirectLoginScreen.tsx
@@ -1,20 +1,20 @@
 import React, { useEffect, useState } from 'react';
 import { useOktaAuth } from '@okta/okta-react';
 import { useTranslation } from 'react-i18next';
-import { OktaLoginScreenBase } from './OktaLoginScreenBase';
-import { OktaLoginScreenProps } from './types';
+import { OktaRedirectLoginScreenBase } from './OktaRedirectLoginScreenBase';
+import { OktaRedirectLoginScreenProps } from './types';
 import { useAuthContext } from '../../contexts';
 import { useErrorManager } from '../../contexts/ErrorContext/useErrorManager';
 
 /**
  * Component that renders an okta login screen.
  *
- * @param {OktaLoginScreenProps} props - props of OktaLoginScreen
+ * @param {OktaRedirectLoginScreenProps} props - props of OktaRedirectLoginScreen
  *
  * @category Component
  */
 
-export const OktaLoginScreen: React.FC<OktaLoginScreenProps> = (props) => {
+export const OktaRedirectLoginScreen: React.FC<OktaRedirectLoginScreenProps> = (props) => {
     const { authState, oktaAuth } = useOktaAuth();
     const [isLoading, setIsLoading] = useState(!authState);
     const { triggerError, errorManagerConfig } = useErrorManager();
@@ -65,7 +65,7 @@ export const OktaLoginScreen: React.FC<OktaLoginScreenProps> = (props) => {
     };
 
     return (
-        <OktaLoginScreenBase
+        <OktaRedirectLoginScreenBase
             loading={isLoading}
             loginButtonLabel={loginButtonLabel}
             onLogin={handleOnLogin}
