+++ login-workflow/src/screens/index.ts
@@ -13,4 +13,4 @@ export * from './ForgotPasswordScreen';
 export * from './ContactScreen';
 export * from './RegistrationSuccessScreen';
 export * from './ExistingAccountSuccessScreen';
-export * from './OktaLoginScreen';
+export * from './OktaRedirectLoginScreen';
